Automatically generate completion file
Wrote a script to use Samtools' usage message to generate all
subcommands. It then uses each of these subcommands' usage messages to
generate long options (this is implemented naively: I simply grep for
'--' in usage messages, omitting options followed by punctuation other
than a dash, as this seems to mean that they are just longer 'short
options'). These subcommands are then used by
'generate_completion_file.sh' to generate a tab completion file. This
file is printed to stdout, and it is up to the user to store this file
in the correct location. I have stored an example of the output file in
'current_samtools_tab_completion_file'.